+++ docs/release_notes.rst
@@ -6,6 +6,8 @@ Release Notes
 -----
 
 * Upgraded Javet to v3.0.3
+* Added module ``javet``
+* Added ``JavetReflectionObjectFactory``
 
 0.3.0
 -----